chore: create TopicEditor and CategoryEditor files
